# Partial fix for benchmarks.