Partial fix for benchmarks.
The thing is that in some microbenchmarks the calculation result is not used in any way, because of this the compiler simply "optimizes the benchmark" by throwing away the payload.

Most of the current benchmark implementations literally measure the air in an empty room.

I can only partially fix them, I think for further fixes you should ask Vyacheslav Egorov for advice.

https://mrale.ph/blog/2021/01/21/microbenchmarking-dart-part-1.html
https://mrale.ph/blog/2024/11/27/microbenchmarks-are-experiments.html


```bash
$ dart run benchmark/vector2_bench.dart
```

|                           | Before                 | After                  |
| ------------------------- | ---------------------- | ---------------------- |
| Vector2()                 | 202.8197194309024 us.  | 3732.437062937063 us.  |
| Vector2.zero()            | 200.17709611451943 us. | 3483.7631184407796 us. |
| Vector2.array()           | 391.3436018957346 us.  | 3571.7739520958085 us. |
| Vector2.all()             | 388.8081180811808 us.  | 3741.715034965035 us.  |
| Vector2.copy()            | 386.95465071401003 us. | 3838.61013986014 us.   |
| Vector2.fromFloat32List() | 389.3649748159628 us.  | 1719.3718140929536 us. |
| Vector2.fromBuffer()      | 5980.412087912088 us.  | 6143.877245508982 us.  |
| Vector2.random()          | 25577.240506329115 us. | 29626.848484848484 us. |
| Vector2.setFrom()         | 243.5085574572127 us.  | 234.90559195133363 us. |
| Vector2.dot()             | 195.64618896387455 us. | 389.8708860759494 us.  |

```bash
$ dart compile exe -o benchmark/vector2_bench.exe benchmark/vector2_bench.dart
$ ./benchmark/vector2_bench.exe
```

|                           | Before                 | After                  |
| ------------------------- | ---------------------- | ---------------------- |
| Vector2()                 | 3862.1923076923076 us. | 3528.176911544228 us.  |
| Vector2.zero()            | 194.03286978508217 us. | 3356.08095952024 us.   |
| Vector2.array()           | 7367.800699300699 us.  | 7313.762237762237 us.  |
| Vector2.all()             | 3860.874125874126 us.  | 3485.455772113943 us.  |
| Vector2.copy()            | 3835.0874125874125 us. | 3624.082 us.           |
| Vector2.fromFloat32List() | 386.48084365325076 us. | 1638.9460269865067 us. |
| Vector2.fromBuffer()      | 6011.754491017964 us.  | 7060.604895104895 us.  |
| Vector2.random()          | 11294.576923076924 us. | 11604.103448275862 us. |
| Vector2.setFrom()         | 519.39625 us.          | 349.6024287222809 us.  |
| Vector2.dot()             | 196.22144129875272 us. | 386.6904991327809 us.  |